style: Product 컴포넌트들 분해 및 파일 이동
